packages/validators/requirements.txt
jsonschema>=4.17.0
pyyaml>=6.0
